Removed test/ and static/ directories
The code being tested in test/ is deprecated, old and unmaintained. We should
clean this mess before inviting people to contribute to SciRuby.

The static/ directory contained only a logo, which I will send as a PR to the
sciruby.com source branch.

And, finally, I removed some things from the gemspec -- too many things there.