Added static/, fixed load path in sciruby-plotter
Reverted the removal of the static/ directory as said in the PR.

The change in bin/sciruby-plotter is necessary for the script to require the
"local" sciruby gem, not the one installed on the system. This way, you can
make changes to your local copy and test then without conflicts with the one
installed, etc.

I didn't add specs because I don't really know how to test a GUI program.